+++ examples/live_output.rb
@@ -1,5 +1,5 @@
 require "./lib/kommando"
 
-Kommando.run "ping -c 3 127.0.0.1", {
+Kommando.run "ping -c 3 -i 0.2 127.0.0.1", {
   output: true
 }